+++ src/models/index.ts
@@ -2,4 +2,5 @@ export * from "./enums";
 export * from "./llm.types";
 export * from "./model.interface";
 export * from "./model.initializer";
+export * from "./model.logic";
 export * from "./rerankers/voyageai-rerank";